web/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
asyncpg
python-dotenv
pydantic-settings
python-multipart
PyJWT
httpx
